+++ src/main/java/com/enn3developer/gtnhvoice/network/NetworkHandler.java
@@ -3,6 +3,8 @@ package com.enn3developer.gtnhvoice.network;
 import com.enn3developer.gtnhvoice.GtnhVoice;
 import com.enn3developer.gtnhvoice.client.ServerHelloClientHandler;
 import com.enn3developer.gtnhvoice.client.ServerRejectClientHandler;
+import com.enn3developer.gtnhvoice.client.VoiceRosterSnapshotClientHandler;
+import com.enn3developer.gtnhvoice.client.VoiceRosterUpdateClientHandler;
 import com.enn3developer.gtnhvoice.server.ClientHelloServerHandler;
 
 import cpw.mods.fml.common.network.NetworkRegistry;
@@ -29,6 +31,12 @@ public final class NetworkHandler {
         WRAPPER.registerMessage(ClientHelloServerHandler.class, ClientHelloPacket.class, id++, Side.SERVER);
         WRAPPER.registerMessage(ServerHelloClientHandler.class, ServerHelloPacket.class, id++, Side.CLIENT);
         WRAPPER.registerMessage(ServerRejectClientHandler.class, ServerRejectPacket.class, id++, Side.CLIENT);
+        WRAPPER.registerMessage(
+            VoiceRosterSnapshotClientHandler.class,
+            VoiceRosterSnapshotPacket.class,
+            id++,
+            Side.CLIENT);
+        WRAPPER.registerMessage(VoiceRosterUpdateClientHandler.class, VoiceRosterUpdatePacket.class, id++, Side.CLIENT);
 
         GtnhVoice.LOG.info(
             "Registered {} control-channel messages on '{}' (hasChannel CLIENT={} SERVER={})",